Frontend serving redesign
This redesign changes how the frontend is handled. It moves the frontend
of the show serving code into its internal/frontend package, which will
contain all components of serving the frontend.

The new server package contains a new instance-based server, which can
be extended using components which are served on the path-prefix. This
Also removes the code responsible for loading individual hosts.