Database interface for users (#24)
* Add multiple database support (#23)

In the future this will allow to use other databases.

* Add user database interface

This change add an interface which can be implemented by various systems
to store the data. This means that we won't have to always rely on SQL.
It is up to implementation to return the correct information.

* Removed the shrug

Ryan didn't want the shrug.